+++ src/Data/ChangePasswordParams.php
@@ -11,6 +11,7 @@ use Upmind\ProvisionBase\Provider\DataSet\Rules;
  * Data used to set a new hosting account password.
  *
  * @property-read string|integer|null $customer_id ID of the customer on the hosting platform
+ * @property-read string|integer|null $subscription_id ID of the subscription on the hosting platform, if any
  * @property-read string $username Username of the account
  * @property-read string $password Desired new password
  */
@@ -20,6 +21,7 @@ class ChangePasswordParams extends DataSet
     {
         return new Rules([
             'customer_id' => ['nullable'],
+            'subscription_id' => ['nullable'],
             'username' => ['required', 'string'],
             'password' => ['required', 'string'],
         ]);