+++ src/Data/GetLoginUrlParams.php
@@ -11,6 +11,7 @@ use Upmind\ProvisionBase\Provider\DataSet\Rules;
  * Data for obtaining a pre-signed login URL.
  *
  * @property-read string|integer|null $customer_id ID of the customer on the hosting platform
+ * @property-read string|integer|null $subscription_id ID of the subscription on the hosting platform, if any
  * @property-read string $username Username of the account
  * @property-read string $user_ip IP of the person who wishes to log in
  * @property-read boolean|null $is_reseller Whether or not the account is a reseller
@@ -21,6 +22,7 @@ class GetLoginUrlParams extends DataSet
     {
         return new Rules([
             'customer_id' => ['nullable'],
+            'subscription_id' => ['nullable'],
             'username' => ['required', 'string'],
             'user_ip' => ['required', 'ip'],
             'is_reseller' => ['boolean'],